+++ lib/gs1.rb
@@ -1,21 +1,3 @@
-require 'gs1/version'
-
-require 'gs1/ai'
-require 'gs1/check_digit_calculator'
-require 'gs1/extensions'
-require 'gs1/definitions'
-require 'gs1/validations'
-
-require 'gs1/record'
-require 'gs1/batch'
-require 'gs1/content'
-require 'gs1/expiration_date'
-require 'gs1/gtin'
-require 'gs1/serial_number'
-require 'gs1/sscc'
-
-require 'gs1/barcode'
-
 # GS1 module.
 #
 module GS1
@@ -48,8 +30,10 @@ module GS1
     end
   end
 
-  AI_CLASSES = GS1::Record.descendants.each_with_object({}) do |klass, hash|
-    hash[klass.ai] = klass
+  def self.ai_classes
+    @ai_classes ||= GS1::Record.descendants.each_with_object({}) do |klass, hash|
+      hash[klass.ai] = klass
+    end
   end
 
   module AIDCMarketingLevels
@@ -58,3 +42,21 @@ module GS1
            HIGHEST = 3].freeze
   end
 end
+
+require 'gs1/version'
+
+require 'gs1/ai'
+require 'gs1/check_digit_calculator'
+require 'gs1/extensions'
+require 'gs1/definitions'
+require 'gs1/validations'
+
+require 'gs1/record'
+require 'gs1/batch'
+require 'gs1/content'
+require 'gs1/expiration_date'
+require 'gs1/gtin'
+require 'gs1/serial_number'
+require 'gs1/sscc'
+
+require 'gs1/barcode'