+++ CHANGELOG.md
@@ -22,8 +22,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added the `GS1.ai_classes` method. This contains all AI classes. This is the
     same as `GS1::AI_CLASSES` from previous versions, buy lazy loaded
 
+- All application identifiers (as of 2025-01-30) are now recognized
+
 ### Removed
 - Removed the `GS1::AI_CLASSES` constant. Use `GS1.ai_classes` instead
+- The following classes have been removed: `GS1::Batch` and `GS1::SerialNumber`.
+    They have been replaced with generated anonymous classes.
 
 ## [1.1.0] - 2021-08-23