+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.0.0] - 2025-04-22
 ### Changed
 - Minimum supported version of Ruby is now 3.0.5 (breaking change)
 
@@ -45,7 +47,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fix `nil` argument for barcode
 
-[Unreleased]: https://github.com/apoex/gs1/compare/v1.1.0...HEAD
+[Unreleased]: https://github.com/apoex/gs1/compare/v2.0.0...HEAD
 
+[2.0.0]: https://github.com/apoex/gs1/compare/v1.1.0...v2.0.0
 [1.1.0]: https://github.com/apoex/gs1/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/apoex/gs1/releases/tag/v1.0.0